Move transaction executor (#19946)
