+++ client/src/lib.rs
@@ -20,3 +20,4 @@ pub mod rpc_response;
 pub mod rpc_sender;
 pub mod thin_client;
 pub mod tpu_client;
+pub mod transaction_executor;
